requirements.txt
streamlit
openai
razorpay
twilio
psycopg2-binary
python-dotenv
pandas
numpy
google-generativeai